Remove legacy (invalid or misleading) chord type
