+++ packages/chord-type/test.ts
@@ -1,6 +1,8 @@
 import ChordType from "./index";
+import { interval } from "@tonaljs/core";
+import DATA from "./data";
 
-const $ = (str: string) => str.split(" ");
+const INTERVALS = DATA.map((d) => d[0]).sort();
 
 describe("@tonaljs/chord-type", () => {
   test("names", () => {
@@ -24,7 +26,7 @@ describe("@tonaljs/chord-type", () => {
   });
 
   test("all returns all chords", () => {
-    expect(ChordType.all()).toHaveLength(107);
+    expect(ChordType.all()).toHaveLength(106);
   });
 
   test("get ", () => {
@@ -53,5 +55,40 @@ describe("@tonaljs/chord-type", () => {
     ChordType.removeAll();
     expect(ChordType.all()).toEqual([]);
     expect(ChordType.keys()).toEqual([]);
+  });
+});
+
+describe("@tonal/chord-type data", () => {
+  test("no repeated intervals", () => {
+    for (let i = 1; i < INTERVALS.length; i++) {
+      expect(INTERVALS[i - 1]).not.toEqual(INTERVALS[i]);
+    }
+  });
+
+  test("all chords must have abreviatures", () => {
+    DATA.forEach((data) => {
+      const abrrvs = data[2].trim();
+      expect(abrrvs.length).toBeGreaterThan(0);
+    });
+  });
+
+  test("intervals should be in ascending order", () => {
+    DATA.forEach((data) => {
+      const [list] = data;
+      const intervals = list.split(" ").map((i) => interval(i).semitones || 0);
+
+      try {
+        for (let i = 1; i < intervals.length; i++) {
+          expect(intervals[i - 1]).toBeLessThan(intervals[i]);
+        }
+      } catch (e) {
+        // tslint:disable-next-line
+        console.error(
+          `Invalid chord: intervals should be in ascending order`,
+          data
+        );
+        throw e;
+      }
+    });
   });
 });
